Overhaul
Starting a new portfolio, Starting with adding in a top nav. I'm having some challenge getting the links to appear once the hamburger is clicked in the menu. I plan on researching how to fix this issue, but I want the links to be animated from the right side of the screen and slide back off screen when the hamburger is closed.