Update - almost everything
I decided to just have 1 main index html file, but create sections. This eliminates the need to update multiple html files with updates to styling and whatnot. I made each section a different color for testing in the present to see if it works and possibly if I like the look of it.  I will have to adjust the mobile link menu for it blends into the main page a bit and doesn't show all of the links...Might contemplate doing away with the topnav.